doc: Refine for FuzzyTokenSet
There were a couple of issues with the existing documentation.
Firstly, clarify that the weights associated with each operation may
be zero (they need not be strictly positive). Secondly, clarify that
the inner comparator may be a distance function or a similarity
function. If the inner comparator is a similarity function, then
the goal is to _maximize_ the cost of operations that transform
the left string into the right string.